Fixed a bug in the pieces available view where, if a piece was created and placed in the same frame, the view would desync with the model. Fixed a bug in the gameplay screen where it was forgetting to call base.Hide thus failing to tell the screen it should hide. Added placeholder art for buttons. Added images, and a background for the instructions.

+++ Strategy/Interface/Screens/MessageScreen.cs
@@ -43,8 +43,8 @@ namespace Strategy.Interface.Screens
             {
                 lineSprites[i] = new TextSprite(font, lines[i]);
                 lineSprites[i].Position = new Vector2(
-                    box.Position.X + (box.Size.X - lineSprites[i].Size.X) / 2,
-                    y + i * font.LineSpacing);
+                    (int)(box.Position.X + (box.Size.X - lineSprites[i].Size.X) / 2),
+                    (int)(y + i * font.LineSpacing));
                 lineSprites[i].Color = new Color(60, 60, 60);
             }
             CompositeSprite message = new CompositeSprite(lineSprites);
